Set parallel lookup table generation as the default
Delete old lookup table generation script
