+++ Dynamics/generate_lookup_table.py
@@ -1,9 +1,21 @@
+"""
+A script to generate a lookup table for the efficiencies of a user-input twobox grating. Generates 
+the efficiencies for a range of wavelengths and angles, and saves the data as a pickle file. The 
+calculation is parallelised using the multiprocess module, so set the number of processes before
+running.
+"""
+
+import itertools
+
+import multiprocess as mp
+
 import numpy as np
 import pickle
-import time
 
 import sys
 sys.path.append('../')
+
+import time
 
 from parameters import D1_ND
 from twobox import TwoBox
@@ -12,16 +24,17 @@ from twobox import TwoBox
 t_start = time.time()
 
 ## Initialise grating
-grating_pitch   = 1.4185910181100811
-grating_depth   = 0.49190319526197407
-box1_width      = 0.5719025530222406
-box2_width      = 0.053439272331534775
-box_centre_dist = 0.44458828885168056
-box1_eps        = 11.029595778447616
-box2_eps        = 6.100136959625866
-gaussian_width  = 31.37144885298504
-substrate_depth = 0.43257336002828756
-substrate_eps   = 2.279546035418172
+# TODO: extract gratings directly from optimised pkl and print parameters
+grating_pitch   = 1.4448460868780606
+grating_depth   = 2.0643298724749126
+box1_width      = 0.7418355927658334
+box2_width      = 0.6154206329992875
+box_centre_dist = 0.32392145077748286
+box1_eps        = 4.56512120699494
+box2_eps        = 4.228855153129463
+gaussian_width  = 14.228774249456437
+substrate_depth = 0.01
+substrate_eps   = 3.7022541961011903
 
 wavelength      = 1.
 angle           = 0.
@@ -34,23 +47,30 @@ grating = TwoBox(grating_pitch, grating_depth, box1_width, box2_width, box_centr
                  wavelength, angle, Nx, numG, Qabs)
 
 
-klambda = 100  # Number of lambda' points
+klambda = 1000  # Number of lambda' points
 v_final = 5/100
 lambda_final = 1/D1_ND(v_final)
 lambda_array = np.linspace(wavelength, lambda_final, klambda)
 
-kdelta = 100  # Number of delta' points
-delta_max = 15*np.pi/180
+kdelta = 1000  # Number of delta' points
+delta_max = 16*np.pi/180
 delta_min = -delta_max
 delta_array = np.linspace(delta_min, delta_max, kdelta)
 
-efficiencies = np.zeros((klambda, kdelta, 6))  # 6 ordered efficiencies: Q1, Q2, PD_Q1_delta, PD_Q2_delta, PD_Q1_lambda, PD_Q2_lambda
+num_processes = 8
 
-for i in range(klambda):
-    grating.wavelength = lambda_array[i]
-    for j in range(kdelta):
-        grating.angle = delta_array[j]
-        efficiencies[i,j,:] = grating.return_Qs_auto()
+def eff_auto(*args):
+    grating.wavelength = args[0][0]
+    grating.angle = args[0][1]
+    return grating.return_Qs_auto() 
+
+param_inputs = ((l,d) for l,d in itertools.product(lambda_array,delta_array))
+
+p = mp.Pool(num_processes)
+efficiencies_ = p.map(eff_auto, param_inputs)
+efficiencies = np.reshape(np.array(efficiencies_), (klambda, kdelta, 6))  # 6 ordered efficiencies: Q1, Q2, PD_Q1_delta, PD_Q2_delta, PD_Q1_lambda, PD_Q2_lambda
+p.close()
+p.join()
 
 Q1_array           = efficiencies[:,:,0];   Q2_array           = efficiencies[:,:,1]
 PD_Q1_delta_array  = efficiencies[:,:,2];   PD_Q2_delta_array  = efficiencies[:,:,3]
@@ -67,5 +87,5 @@ pkl_fname = rf"./Data/Lookup_table_lambda_{klambda}_by_delta_{kdelta}.pkl"
 data = {'Q1': Q1_array, 'Q2': Q2_array, 'PD_Q1_delta': PD_Q1_delta_array, 'PD_Q2_delta': PD_Q2_delta_array, 
         'PD_Q1_lambda': PD_Q1_lambda_array, 'PD_Q2_lambda': PD_Q2_lambda_array, 
         'lambda array': lambda_array, 'delta array': delta_array}
-# with open(pkl_fname, 'wb') as data_file:
-#             pickle.dump(data, data_file)
+with open(pkl_fname, 'wb') as data_file:
+    pickle.dump(data, data_file)
